requirements.txt
django==5.0.6
asgiref==3.8.1
tzdata==2024.1
sqlparse==0.4.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pillow==10.2.0
python-dotenv==1.0.0
django-filter==24.2
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
coverage==7.0.5
drf-yasg==1.21.7
inflection==0.5.1
PyYAML==6.0.1
stripe==9.9.0
celery==5.4.0
billiard==4.2.0
click==8.1.7
click_didyoumean==0.3.1
click_plugins==1.1.1
click_repl==0.3.0
kombu==5.3.7
python_dateutil==2.9.0.post0
vine==5.1.0
amqp==5.2.0
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
redis==4.5.5
cron-descriptor==1.4.3
django-timezone-field==6.1.0
django_celery_beat==2.6.0
python-crontab==3.1.0
